move buffered reader to its own header
